# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "text": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResult": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/PlanCard"
            },
            "title": "Cards",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "cards",
          "message"
        ],
        "title": "AskResult",
        "type": "object"
      },
...
      "BindRequest": {
        "properties": {
          "all_or_nothing": {
            "default": false,
            "title": "All Or Nothing",
            "type": "boolean"
          },
          "quote_id": {
            "title": "Quote Id",
            "type": "string"
          }
        },
        "required": [
          "quote_id"
        ],
        "title": "BindRequest",
        "type": "object"
      },
      "Business": {
        "properties": {
          "bad_day_dollars": {
            "title": "Bad Day Dollars",
            "type": "integer"
          },
          "bank": {
            "$ref": "#/components/schemas/BankLink"
          },
          "city": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/City"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "topics": {
            "items": {
              "enum": [
                "weather",
                "fuel",
                "rates",
                "prices",
                "tariffs",
                "sports",
                "jobs",
                "other"
              ],
              "type": "string"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "industry",
          "topics",
          "bad_day_dollars",
          "bank",
          "created_at"
        ],
        "title": "Business",
        "type": "object"
      },
      "BusinessListItem": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/City"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "industry"
        ],
        "title": "BusinessListItem",
        "type": "object"
      },
...
      "City": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "state"
        ],
        "title": "City",
        "type": "object"
      },
      "CreateBusiness": {
        "properties": {
          "city_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "description": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "CreateBusiness",
        "type": "object"
      },
      "Forecast": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/PlanCard"
            },
            "title": "Cards",
            "type": "array"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "tailored": {
            "title": "Tailored",
            "type": "boolean"
          }
        },
        "required": [
          "cards",
          "tailored"
        ],
        "title": "Forecast",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyDetail": {
        "properties": {
          "catch": {
            "title": "Catch",
            "type": "string"
          },
          "closes_at": {
            "title": "Closes At",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/PolicyEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/PolicyLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "max_payout_cents": {
            "title": "Max Payout Cents",
            "type": "integer"
          },
          "movements": {
            "items": {
              "$ref": "#/components/schemas/MoneyMovement"
            },
            "title": "Movements",
            "type": "array"
          },
          "paid_cents": {
            "title": "Paid Cents",
            "type": "integer"
          },
          "payout_each_cents": {
            "title": "Payout Each Cents",
            "type": "integer"
          },
          "premium_cents": {
            "title": "Premium Cents",
            "type": "integer"
          },
          "status": {
            "enum": [
              "PENDING",
              "ACTIVE",
              "AWAITING_RESULT",
              "PAID",
              "EXPIRED",
              "REFUNDED",
              "NEEDS_REVIEW"
            ],
            "title": "Status",
            "type": "string"
          },
          "terms": {
            "title": "Terms",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic": {
            "enum": [
              "weather",
              "fuel",
              "rates",
              "prices",
              "tariffs",
              "sports",
              "jobs",
              "other"
            ],
            "title": "Topic",
            "type": "string"
          },
          "why": {
            "title": "Why",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic",
          "title",
          "status",
          "closes_at",
          "payout_each_cents",
          "max_payout_cents",
          "premium_cents",
          "paid_cents",
          "created_at",
          "why",
          "catch",
          "terms",
          "legs",
          "events",
          "movements"
        ],
        "title": "PolicyDetail",
        "type": "object"
      },
      "PolicyEvent": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message",
          "created_at"
        ],
        "title": "PolicyEvent",
        "type": "object"
      },
...
      "PolicySummary": {
        "properties": {
          "closes_at": {
            "title": "Closes At",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "max_payout_cents": {
            "title": "Max Payout Cents",
            "type": "integer"
          },
          "paid_cents": {
            "title": "Paid Cents",
            "type": "integer"
          },
          "payout_each_cents": {
            "title": "Payout Each Cents",
            "type": "integer"
          },
          "premium_cents": {
            "title": "Premium Cents",
            "type": "integer"
          },
          "status": {
            "enum": [
              "PENDING",
              "ACTIVE",
              "AWAITING_RESULT",
              "PAID",
              "EXPIRED",
              "REFUNDED",
              "NEEDS_REVIEW"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic": {
            "enum": [
              "weather",
              "fuel",
              "rates",
              "prices",
              "tariffs",
              "sports",
              "jobs",
              "other"
            ],
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic",
          "title",
          "status",
          "closes_at",
          "payout_each_cents",
          "max_payout_cents",
          "premium_cents",
          "paid_cents",
          "created_at"
        ],
        "title": "PolicySummary",
        "type": "object"
      },
      "Quote": {
        "properties": {
          "breakdown": {
            "$ref": "#/components/schemas/QuoteBreakdown"
          },
          "catch": {
            "title": "Catch",
            "type": "string"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/QuoteLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "max_payout_cents": {
            "title": "Max Payout Cents",
            "type": "integer"
          },
          "payout_each_cents": {
            "title": "Payout Each Cents",
            "type": "integer"
          },
          "premium_cents": {
            "title": "Premium Cents",
            "type": "integer"
          },
          "terms": {
            "title": "Terms",
            "type": "string"
          },
          "thin_book": {
            "$ref": "#/components/schemas/ThinBook"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic": {
            "enum": [
              "weather",
              "fuel",
              "rates",
              "prices",
              "tariffs",
              "sports",
              "jobs",
              "other"
            ],
            "title": "Topic",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "why": {
            "title": "Why",
            "type": "string"
          }
        },
        "required": [
          "id",
          "expires_at",
          "title",
          "topic",
          "why",
          "catch",
          "payout_each_cents",
          "max_payout_cents",
          "premium_cents",
          "legs",
          "breakdown",
          "terms",
          "warnings",
          "thin_book"
        ],
        "title": "Quote",
        "type": "object"
      },
...
      "QuoteRequest": {
        "properties": {
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegRequest"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Legs",
            "type": "array"
          },
          "payout_dollars": {
            "minimum": 1.0,
            "title": "Payout Dollars",
            "type": "integer"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanText"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "legs",
          "payout_dollars"
        ],
        "title": "QuoteRequest",
        "type": "object"
      },
      "ResolveRequest": {
        "properties": {
          "result": {
            "enum": [
              "yes",
              "no"
            ],
            "title": "Result",
            "type": "string"
          }
        },
        "required": [
          "result"
        ],
        "title": "ResolveRequest",
        "type": "object"
      },
      "SessionRequest": {
        "properties": {
          "business_id": {
            "title": "Business Id",
            "type": "integer"
          }
        },
        "required": [
          "business_id"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
...
      "TopicsUpdate": {
        "properties": {
          "topics": {
            "items": {
              "enum": [
                "weather",
                "fuel",
                "rates",
                "prices",
                "tariffs",
                "sports",
                "jobs",
                "other"
              ],
              "type": "string"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "topics"
        ],
        "title": "TopicsUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HedgeCast",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ask": {
      "post": {
        "operationId": "ask_api_ask_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask"
      }
    },
    "/api/businesses": {
      "get": {
        "operationId": "businesses_api_businesses_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BusinessListItem"
                  },
                  "title": "Response Businesses Api Businesses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Businesses"
      },
      "post": {
        "operationId": "create_business_api_businesses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusiness"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Business"
      }
    },
...
    "/api/forecast": {
      "get": {
        "operationId": "forecast_api_forecast_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forecast"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forecast"
      }
    },
...
    "/api/me": {
      "get": {
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me"
      }
    },
    "/api/me/bank": {
      "post": {
        "operationId": "link_bank_api_me_bank_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link Bank"
      }
    },
    "/api/me/topics": {
      "put": {
        "operationId": "set_topics_api_me_topics_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Topics"
      }
    },
...
    "/api/ops/policies/{policy_id}/resolve": {
      "post": {
        "operationId": "ops_resolve_api_ops_policies__policy_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ops Resolve"
      }
    },
...
    "/api/policies": {
      "get": {
        "operationId": "policies_api_policies_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicySummary"
                  },
                  "title": "Response Policies Api Policies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Policies"
      },
      "post": {
        "operationId": "bind_api_policies_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bind"
      }
    },
    "/api/policies/{policy_id}": {
      "get": {
        "operationId": "policy_api_policies__policy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Policy"
      }
    },
    "/api/quotes": {
      "post": {
        "operationId": "quote_api_quotes_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "hc_business",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Business"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quote"
      }
    },
    "/api/session": {
      "delete": {
        "operationId": "end_session_api_session_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "End Session"
      },
      "post": {
        "operationId": "start_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session"
      }
    },
...
  }
}